+++ KWEngine3/Helper/SerializedRenderObject.cs
@@ -8,7 +8,7 @@ namespace KWEngine3.Helper
     {
         // MISC. PROPERTIES
         public bool IsShadowCaster { get; set; }
-        public int ID { get; set; }
+        public ushort ID { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
         public string ModelName { get; set; }
